requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.1
defusedxml==0.7.1
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
django-ipware==7.0.1
django-redis==5.4.0
django-rest-authemail==2.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
httpie==3.2.2
idna==3.7
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
Pygments==2.18.0
PyJWT==1.7.1
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
python-http-client==3.3.7
python-ipware==3.0.0
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
sendgrid==6.11.0
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
str2bool==1.1
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.6.0
yarl==1.9.4
